prod_cmp/requirements.txt
langchain-mcp-adapters
langgraph
langchain-openai
langchain
python-dotenv
mcp
streamlit
fastapi
uvicorn
langchain-community
google-search-results